requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
dictdiffer==0.8.0
google-auth==1.6.3
idna==2.8
Jinja2==2.10.1
Flask==1.1.1
kubernetes==9.0.0
MarkupSafe==1.1.1
oauthlib==3.0.2
openshift==0.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
python-string-utils==0.6.0
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel.yaml==0.15.99
six==1.12.0
setuptools==21.0.0
urllib3==1.25.3
websocket_client==0.56.0
